refactor: share editor service types (#85)
