+++ src/editor/bbpress-adapter.ts
@@ -3,6 +3,11 @@
  */
 import { addFilter } from '@wordpress/hooks';
 
+/**
+ * Internal dependencies
+ */
+import type { EditorServiceContext, EditorServices } from './editor-services';
+
 type BbPressAdapterOptions = {
 	container: HTMLElement;
 	settings: typeof wpBlocksEverywhere;
@@ -17,49 +22,6 @@ type BbPressAdapterOptions = {
 		context: EditorServiceContext,
 		details?: unknown
 	) => void;
-};
-
-type EditorServiceContext = {
-	container?: HTMLElement;
-	editorType?: string;
-	mode?: 'inline' | 'full-height' | 'modal' | 'compact';
-	settings: typeof wpBlocksEverywhere;
-	textarea?: HTMLTextAreaElement;
-};
-
-type EditorAutosaveService =
-	| ( ( payload: Record< string, unknown >, context: EditorServiceContext ) => unknown )
-	| {
-			delay?: number;
-			save?: ( payload: Record< string, unknown >, context: EditorServiceContext ) => unknown;
-			cancel?: ( context: EditorServiceContext ) => void;
-	  }
-	| null;
-
-type EditorNoticeService =
-	| ( ( type: string, message: string, context: EditorServiceContext, details?: unknown ) => void )
-	| {
-			error?: ( message: string, context: EditorServiceContext, details?: unknown ) => void;
-			success?: ( message: string, context: EditorServiceContext, details?: unknown ) => void;
-			warning?: ( message: string, context: EditorServiceContext, details?: unknown ) => void;
-			info?: ( message: string, context: EditorServiceContext, details?: unknown ) => void;
-	  }
-	| null;
-
-type EditorPermissionsService = {
-	can?: ( capability: string, context: EditorServiceContext ) => boolean | undefined;
-	canUploadMedia?: boolean | ( ( context: EditorServiceContext ) => boolean | undefined );
-} | null;
-
-type EditorServices = {
-	apiFetch?: ( options: Record< string, unknown > ) => Promise< unknown >;
-	apiFetchMiddleware?: ( options: Record< string, unknown >, next: Function ) => unknown;
-	apiFetchMiddlewares?: Array< ( options: Record< string, unknown >, next: Function ) => unknown >;
-	autosave?: EditorAutosaveService;
-	fetchLinkSuggestions?: ( search: string, searchOptions?: Record< string, unknown > ) => Promise< unknown >;
-	mediaUpload?: Function | null;
-	notices?: EditorNoticeService;
-	permissions?: EditorPermissionsService;
 };
 
 let hasInstalledAutocompleteCompatibilityFilter = false;
